refactor: Split JSON Schema conversion to separate function
